--Empty Message--

+++ blmarket/TriangleXor.cpp
@@ -43,9 +43,9 @@ public:
             if(nc <= 0) break;
             double y1 = (double)i / (W + i);
             double y0 = (double)(i-1) / (W + i-1);
-            double xspace = W / (W+(i+1)/2);
+            double xspace = (double)W / (W+(i+1)/2);
             cout << y1 << " " << y0 << " " << nc << " " << xspace << endl;
-            ret += (y1-y0) / 2 * nc * W / (W + (i+1)/2);
+            ret += (y1-y0) / 2.0 * nc * W / (W + (i+1)/2);
             if((i%2) == 0) nc -= 2;
         }
